refactor(plan-detail): move ProjectPlanDetailPage into plan-detail/
