+++ frontend/src/react/pages/project/plan-detail/index.ts
@@ -1,0 +1,1 @@
+export { ProjectPlanDetailPage } from "./ProjectPlanDetailPage";
